Added 'UserView' user profile view, code review, and numerous improvements

+++ myShop-angular/src/app/components/user-cart/user-cart.ts
@@ -3,6 +3,7 @@ import { FormsModule } from '@angular/forms';
 import { Router } from '@angular/router';
 import { CartService } from '../../services/cart-service';
 import { AuthService } from '../../services/auth-service';
+import { Common } from '../../constants/common';
 import { Cart } from '../../models/cart';
 import { Product } from '../../models/product';
 import { User } from '../../models/user';
@@ -14,43 +15,52 @@ import { User } from '../../models/user';
   styleUrl: './user-cart.css',
 })
 export class UserCart {
+  // Constants
+  public Common = Common;
+
+  // Native classes / Application services
+  private router = inject(Router);
   private cartService = inject(CartService);
   private authService = inject(AuthService);
-  private router = inject(Router);
 
+  // User messages
+  private static msgDelProduct: string =
+    'Êtes-vous sûr de vouloir retirer cet article de votre panier ?';
+  private static msgDelCart: string = 'Êtes-vous sûr de vouloir supprimer votre panier ?';
+  private static msgConnectToOrder: string =
+    'Vous devez être connecté pour passer commande.\nSouhaitez-vous vous connecter ou vous inscrire ?';
+
+  // Class properties
   tax: number = 1.2; // +20%
   total: number = 0;
   userCart!: Cart;
-  productsIni!: Product[];
-  cart!: string;
   connectedUser: User | null = null;
 
-  // Initialize the Cart object and the view (template)
+  // Initialize the Cart object and the view (template) ///////////////////////
+
   ngOnInit() {
     // Get the user if he is logged in
     this.getConnectedUser();
 
     const CART = localStorage.getItem('cart');
     if (CART) {
-      this.cart = CART;
       const USER_CART = JSON.parse(CART).map((item: any) => {
         const PRODUCT = new Product();
         Object.assign(PRODUCT, item);
         return PRODUCT;
       });
       this.userCart = new Cart(USER_CART);
-      this.productsIni = structuredClone(this.userCart.products);
     } else {
       this.userCart = new Cart();
     }
   }
 
-  // Product-related action ///////////////////////////////////////////////////
-
   // Method to retrieve the logged-in user
   getConnectedUser() {
     this.connectedUser = this.authService.getConnectedUser();
   }
+
+  // Product-related action ///////////////////////////////////////////////////
 
   // View a product
   view(id: string) {
@@ -59,19 +69,18 @@ export class UserCart {
 
   // Increase the quantity of a product
   addOne(product: Product) {
-    // Add the product to local storage via CartService
-    const PRODUCT = new Product();
-    Object.assign(PRODUCT, product);
-    PRODUCT.cartQuantity = 1;
-    this.cartService.add(PRODUCT);
+    // Add an item to the cart via CartService (local storage)
+    this.cartService.add(product);
 
     // Update the cart view
-    product.cartQuantity++;
+    if (typeof product.cartQuantity === 'number') product.cartQuantity++;
   }
 
   // Decrease the quantity of a product
   removeOne(product: Product) {
-    // Remove the product or decrease its quantity in local storage
+    if (product.cartQuantity === 1 && !confirm(UserCart.msgDelProduct)) return;
+
+    // Remove an item from the cart via CartService (local storage)
     this.cartService.remove(product);
 
     // Update the cart view
@@ -79,14 +88,16 @@ export class UserCart {
       this.userCart.products = this.userCart.products.filter(
         (item: Product) => item.id !== product.id,
       );
-    } else {
+    } else if (typeof product.cartQuantity === 'number') {
       product.cartQuantity--;
     }
   }
 
   // Remove a product
   removeProduct(id: string) {
-    // Remove the product from the local storage
+    if (!confirm(UserCart.msgDelProduct)) return;
+
+    // Remove the product from the cart via CartService (local storage)
     this.cartService.removeProduct(id);
 
     // Remove the product from the cart view
@@ -98,28 +109,29 @@ export class UserCart {
   // Retrieve the total cart amount excluding tax
   getTotalExcludingTax() {
     return (this.total = this.userCart.products.reduce((total: number, product: Product) => {
-      return total + product.cartQuantity * product.price;
+      return (
+        total +
+        (typeof product.cartQuantity === 'number' ? product.cartQuantity * product.price : 0)
+      );
     }, 0));
   }
 
-  // Remove the cart
+  // Delete the user's cart
   removeCart() {
-    // Remove the cart from the local storage
+    if (!confirm(UserCart.msgDelCart)) return;
+
+    // Delete the user's cart via CartService (local storage)
     this.cartService.removeCart();
 
     // Remove the cart from the view
     this.userCart.products = [];
   }
 
-  // Place the order
+  // To order
   addOrder() {
     if (this.connectedUser) {
       this.router.navigate(['/add-order', this.connectedUser.id]);
-    } else if (
-      confirm(
-        'Vous devez être connecté pour passer commande.\nSouhaitez-vous vous connecter ou vous inscrire ?',
-      )
-    ) {
+    } else if (confirm(UserCart.msgConnectToOrder)) {
       this.router.navigate(['/login-cart']);
     }
   }
